+++ src/Type/ChatAdministratorRights.php
@@ -27,6 +27,7 @@ final readonly class ChatAdministratorRights
         public ?bool $canEditMessages = null,
         public ?bool $canPinMessages = null,
         public ?bool $canManageTopics = null,
+        public ?bool $canManageDirectMessages = null,
     ) {}
 
     public function toRequestArray(): array
@@ -48,6 +49,7 @@ final readonly class ChatAdministratorRights
                 'can_edit_messages' => $this->canEditMessages,
                 'can_pin_messages' => $this->canPinMessages,
                 'can_manage_topics' => $this->canManageTopics,
+                'can_manage_direct_messages' => $this->canManageDirectMessages,
             ],
             static fn(mixed $value): bool => $value !== null,
         );