+++ tests/Type/ChatAdministratorRightsTest.php
@@ -34,6 +34,7 @@ final class ChatAdministratorRightsTest extends TestCase
         assertNull($rights->canEditMessages);
         assertNull($rights->canPinMessages);
         assertNull($rights->canManageTopics);
+        assertNull($rights->canManageDirectMessages);
 
         assertSame(
             [
@@ -71,6 +72,7 @@ final class ChatAdministratorRightsTest extends TestCase
             'can_edit_messages' => true,
             'can_pin_messages' => false,
             'can_manage_topics' => true,
+            'can_manage_direct_messages' => false,
         ], null, ChatAdministratorRights::class);
 
         assertTrue($type->isAnonymous);
@@ -88,5 +90,27 @@ final class ChatAdministratorRightsTest extends TestCase
         assertTrue($type->canEditMessages);
         assertFalse($type->canPinMessages);
         assertTrue($type->canManageTopics);
+        assertFalse($type->canManageDirectMessages);
+        assertSame(
+            [
+                'is_anonymous' => true,
+                'can_manage_chat' => false,
+                'can_delete_messages' => true,
+                'can_manage_video_chats' => true,
+                'can_restrict_members' => false,
+                'can_promote_members' => true,
+                'can_change_info' => true,
+                'can_invite_users' => true,
+                'can_post_stories' => true,
+                'can_edit_stories' => true,
+                'can_delete_stories' => false,
+                'can_post_messages' => true,
+                'can_edit_messages' => true,
+                'can_pin_messages' => false,
+                'can_manage_topics' => true,
+                'can_manage_direct_messages' => false,
+            ],
+            $type->toRequestArray(),
+        );
     }
 }